Update: Use miragejs graphql library rather than ember-cli-mirage-graphql
