+++ packages/data/addon/mirage/factories/async-query-result.js
@@ -1,0 +1,7 @@
+/**
+ * Copyright 2020, Yahoo Holdings Inc.
+ * Licensed under the terms of the MIT license. See accompanying LICENSE.md file for terms.
+ */
+import { Factory } from 'ember-cli-mirage';
+
+export default Factory.extend();
