+++ packages/data/package.json
@@ -28,6 +28,7 @@
     "postpublish": "ember ts:clean"
   },
   "dependencies": {
+    "@miragejs/graphql": "^0.1.8",
     "apollo-link-context": "^1.0.19",
     "broccoli-funnel": "^3.0.1",
     "broccoli-merge-trees": "^3.0.2",
@@ -37,7 +38,6 @@
     "ember-cli-babel": "^7.20.5",
     "ember-cli-htmlbars": "^5.1.2",
     "ember-cli-mirage": "~1.1.6",
-    "ember-cli-mirage-graphql": "^0.3.3",
     "ember-cli-moment-shim": "^3.8.0",
     "ember-cli-typescript": "^4.0.0",
     "ember-get-config": "0.2.4",
@@ -47,6 +47,7 @@
     "graphql": "^14.6.0",
     "graphql-tag": "^2.10.1",
     "lodash-es": "^4.17.11",
+    "miragejs": "^0.1.41",
     "papaparse": "^5.1.1",
     "pretender": "^3.4.3"
   },
